+++ Changes.md
@@ -1,11 +1,15 @@
 1.3.x.x (relative to 1.3.5.0)
 =======
 
+Features
+--------
+
+- LightTool : Added manipulator for disk and point light radii.
+
 Fixes
 -----
 
 - Windows : Fixed a bug preventing anything except strings from being copied and pasted.
-
 
 1.3.5.0 (relative to 1.3.4.0)
 =======